Improve security in views.py

+++ articles/views.py
@@ -50,8 +50,9 @@ def edit_article(request, pk=None):
     form = ArticlesForm(request.POST or None, instance=article)
     if form.is_valid():
         article = form.save(commit=False)
-        article.updated_at = timezone.now()
-        article.save()
+        if request.user==article.author:
+            article.updated_at = timezone.now()
+            article.save()
         return redirect('articles:blog')
 
     context = {'title':article.title, 'article':article, 'form':form}
@@ -60,7 +61,8 @@ def edit_article(request, pk=None):
 @login_required(login_url="/users/login/")
 def delete_article(request, pk=None):
     article = get_object_or_404(Articles, pk=pk)
-    article.delete()
+    if request.user == article.author:
+        article.delete()
     return redirect('articles:blog')
